+++ pkg/cli/admin/nodeimage/monitor.go
@@ -1,4 +1,4 @@
-package addnodes
+package nodeimage
 
 import (
 	"context"
